refactor(articles): split the articles controller by route group
`check:runtime-complexity` flagged articles.controller.ts at 890 physical
lines against the 500-line controller ceiling. The preview-link route
brought the file into the changed-file scan, so the pre-existing bloat
now has to be resolved rather than carried.

Split along cohesive route groups, mirroring how other large collections
decompose (posts, images, videos):

- `operations/ArticlesOperationsController` — generation and review
  (`POST /articles/generations`, `POST /articles/:articleId/reviews`)
  plus the credit-overdraft preflight those two share.
- `transformations/ArticlesTransformationsController` — the eight
  article-to-artifact routes (thread conversions, virality analyses,
  enhancements, SEO scores, remixes, prompts, images, videos).
- `ArticlesController` keeps CRUD, versions, and the new preview-link
  route, and drops to 246 lines.

Route paths, guards, interceptors, DTOs, and error payloads are
unchanged; only the owning class moves. Both new controllers register
ahead of `ArticlesController` in the module because their static path
segments would otherwise be shadowed by the inherited BaseCRUD `:id`
routes — the same ordering constraint documented in posts.module.ts.

Specs follow their routes, and `ArticlesController`'s spec now provides
the ConfigService that preview-link minting depends on.

openapi.json is updated for the ten operationId renames (operationIds
are `<ControllerClass>.<method>`) and the new preview-links path.
